feat: 유저 즐겨찾기 지역 삭제 api 구현

+++ src/main/java/com/weady/weady/domain/user/repository/UserFavoriteLocationRepository.java
@@ -5,7 +5,7 @@ import com.weady.weady.domain.user.entity.User;
 import com.weady.weady.domain.user.entity.UserFavoriteLocation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserFavoriteLocationRepository extends JpaRepository<UserFavoriteLocation, Integer> {
+public interface UserFavoriteLocationRepository extends JpaRepository<UserFavoriteLocation, Long> {
 
     //지역 중복체크
     Boolean existsByUserAndLocation(User user, Location location);
